+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Table of Contents
 
+- [v1.4.1](#v141)
 - [v1.4.0](#v140)
 - [v1.3.0](#v130)
 - [v1.2.3](#v123)
@@ -24,16 +25,24 @@
 
 ## Unreleased
 
+## [v1.4.1]
+
+> Release date: 2024-11-28
+
 ### Fixes
 
 - Fix setting the `ServiceAccountName` for `DataPlane`'s `Deployment`.
-  [#856](https://github.com/Kong/gateway-operator/pull/856)
+  [#897](https://github.com/Kong/gateway-operator/pull/897)
 - Fixed setting `ExternalTrafficPolicy` on `DataPlane`'s ingress `Service` when
   the requested value is empty.
-  [#865](https://github.com/Kong/gateway-operator/pull/865)
+  [#898](https://github.com/Kong/gateway-operator/pull/898)
+- Set 0 members on `KonnectGatewayControlPlane` which type is set to group.
+  [#896](https://github.com/Kong/gateway-operator/pull/896)
 - Fixed a `panic` in `KonnectAPIAuthConfigurationReconciler` occuring when nil
   response was returned by Konnect API when fetching the organization information.
-  [#890](https://github.com/Kong/gateway-operator/pull/890)
+  [#901](https://github.com/Kong/gateway-operator/pull/901)
+- Bump sdk-konnect-go version to 0.1.10 to fix handling global API endpoints.
+  [#894](https://github.com/Kong/gateway-operator/pull/894)
 
 ## [v1.4.0]
 
@@ -806,6 +815,7 @@ leftovers from previous operator deployments in the cluster. The user needs to d
 (clusterrole, clusterrolebinding, validatingWebhookConfiguration) before
 re-installing the operator through the bundle.
 
+[v1.4.1]: https://github.com/Kong/gateway-operator/compare/v1.4.0..v1.4.1
 [v1.4.0]: https://github.com/Kong/gateway-operator/compare/v1.3.0..v1.4.0
 [v1.3.0]: https://github.com/Kong/gateway-operator/compare/v1.2.3..v1.3.0
 [v1.2.3]: https://github.com/Kong/gateway-operator/compare/v1.2.2..v1.2.3